+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+* Add a configuration for the photo albums navigation bar shadow color [#9](https://github.com/carousell/pickle/pull/9)
 * Fix the photo albums layout after device rotation [#9](https://github.com/carousell/pickle/pull/9)
 
 ## v1.0.0